update so invest page calculates correctly
