+++ modules/pool/lib/usePoolUserTokenBalancesInWallet.tsx
@@ -1,5 +1,5 @@
 import { useUserBalances } from '~/lib/user/useUserBalances';
-import { sumBy } from 'lodash';
+import { sumBy, orderBy } from 'lodash';
 import { useGetTokens } from '~/lib/global/useToken';
 import { AmountHumanReadable } from '~/lib/services/token/token-types';
 import { usePool } from '~/modules/pool/lib/usePool';
@@ -11,8 +11,21 @@ export function _usePoolUserTokenBalancesInWallet() {
 
     const { userBalances, getUserBalance, ...userBalancesQuery } = useUserBalances(allTokenAddresses, allTokens);
 
-    const investTokens = pool.investConfig.options.map((option) => option.tokenOptions).flat();
-    const investableAmount = sumBy(investTokens, (token) =>
+    const tokenOptionsWithHighestValue = pool.investConfig.options.map((option) => {
+        const tokenWithHighestValue = orderBy(
+            option.tokenOptions,
+            (tokenOption) => priceForAmount({ ...tokenOption, amount: getUserBalanceForToken(tokenOption.address) }),
+            'desc',
+        )[0].address;
+
+        return {
+            address: tokenWithHighestValue,
+            amount: getUserBalanceForToken(tokenWithHighestValue),
+            hasMultipleTokenOptions: option.tokenOptions.length > 1,
+        };
+    });
+
+    const investableAmount = sumBy(tokenOptionsWithHighestValue, (token) =>
         priceForAmount({ address: token.address, amount: getUserBalance(token.address) }),
     );
 
@@ -25,6 +38,7 @@ export function _usePoolUserTokenBalancesInWallet() {
         userPoolTokenBalances: userBalances,
         investableAmount,
         getUserBalanceForToken,
+        tokenOptionsWithHighestValue,
     };
 }
